+++ src/main/java/com/Ultra_Nerd/CodeLyokoLegacy/Util/client/sky/volcano/CustomVolcanoSky.java
@@ -35,7 +35,7 @@ public record CustomVolcanoSky() implements DimensionRenderingRegistry.SkyRender
         final Tessellator tessellator = Tessellator.getInstance();
         final BufferBuilder bufferBuilder = tessellator.getBuffer();
         final MatrixStack matrixStack = ctx.matrixStack();
-        for(int i = 0; i < 6; ++i) {
+        for(byte i = 0; i < 6; ++i) {
             matrixStack.push();
             switch (i)
             {
